+++ src/client/js/utils/validateInput.js
